pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "triviagamel6ic"
version = "0.1.0"
[tool.poetry.dependencies]
flask = "^1.1"
python = "^3.7"
mysql-connector = "^2.2"
django-question = "^1.2"
[tool.poetry.dev-dependencies]
